Getting bowhead setup running.
I do not want to git repo to checkout. It will overwrite stuff when I
have a volume mount. I exposes port 8000 because I will be running the
web server in the container as well.

For those who dont know: `php artisan serve --host=0.0.0.0` inside the
container (using screen or whatever you want). Host is needed because
when browsing from the docker host the IP address will not be 127.0.0.1
but one ip on the docker subnet range.

Migration fix: I used 255 everywhere.